requirements.txt
pyyaml
speedtest-cli
requests
fire
aiohttp
tweepy